Created new Route class to replace old one Created shuttle bus helper script Created shuttle bus button in the directions view layout Created location metadata class to consolidate the littered start/end startName/endName Started bringing shuttlebus logic into this branch from original branch
