+++ app/src/main/java/com/example/campusguide/directions/LocationMetadata.kt
@@ -1,0 +1,8 @@
+package com.example.campusguide.directions
+
+data class LocationMetadata constructor(
+    public val name: String,
+    public val encoded: String
+) {
+
+}
