+++ app/src/main/java/com/example/campusguide/directions/Route.kt
@@ -1,171 +1,78 @@
-package com.example.campusguide.directions
+package com.example.campusguide.directions;
 
-import android.widget.RadioButton
-import androidx.fragment.app.FragmentActivity
+import android.location.Location
+import com.example.campusguide.Constants
+import com.example.campusguide.directions.indoor.IndoorSegment
 import com.example.campusguide.directions.outdoor.OutdoorDirections
-import com.example.campusguide.map.Map
-import com.example.campusguide.map.Marker
-import com.example.campusguide.utils.DisplayMessageErrorListener
-import com.example.campusguide.utils.Helper
-import com.example.campusguide.utils.request.ApiKeyRequestDecorator
-import com.example.campusguide.utils.request.VolleyRequestDispatcher
-import com.google.android.gms.maps.CameraUpdateFactory
-import com.google.android.gms.maps.model.Dash
-import com.google.android.gms.maps.model.Gap
-import com.google.android.gms.maps.model.LatLng
-import com.google.android.gms.maps.model.LatLngBounds
-import com.google.android.gms.maps.model.MarkerOptions
-import com.google.android.gms.maps.model.PatternItem
-import com.google.android.gms.maps.model.Polyline
-import com.google.android.gms.maps.model.PolylineOptions
-import com.google.maps.internal.PolylineEncoding
-import kotlinx.coroutines.GlobalScope
-import kotlinx.coroutines.launch
+import com.example.campusguide.directions.outdoor.OutdoorSegment
+import com.example.campusguide.search.indoor.BuildingIndexSingleton
 
-/**
- * Represents a route between two coordinates on the map.
- * @param map The google map object on which the route will be displayed
- * @param ctx The context of the activity that contains the map
- * @param start The name of the location where the route starts
- * @param end The name of the location where the route ends
- */
-class Route constructor(private val map: Map, private val activity: FragmentActivity) {
-    private val PATTERN_DASH_LENGTH_PX = 20
-    private val PATTERN_GAP_LENGTH_PX = 20
-    private val DASH: PatternItem = Dash(PATTERN_DASH_LENGTH_PX.toFloat())
-    private val GAP: PatternItem = Gap(PATTERN_GAP_LENGTH_PX.toFloat())
-    private val PATTERN_POLYGON_ALPHA = listOf(GAP, DASH)
-    private val COLOR_BLUE_ARGB = 0xff0000ff
-    private var polyline: Polyline? = null
-    private var begin: Marker? = null
-    private var dest: Marker? = null
-    private val startString: String = "Start"
-    private val destString: String = "Destination"
+public class Route constructor (
+    private val start: LocationMetadata,
+    private val end: LocationMetadata,
+    private val travelMode: String,
+    private val buildingIndexSingleton: BuildingIndexSingleton,
+    private val giveMeAnOutdoorDirections: () -> OutdoorDirections
+) {
+    private lateinit var pathPolylines: MutableList<PathPolyline>
 
-    fun set(start: String, end: String, travelMode: String) {
-        polyline?.remove()
-        begin?.remove()
-        dest?.remove()
+    init {
+        when(travelMode) {
+            "shuttle" -> {
 
-        val errorListener = DisplayMessageErrorListener(activity)
-        val directions = OutdoorDirections(
-            ApiKeyRequestDecorator(
-                activity,
-                VolleyRequestDispatcher(
-                    activity,
-                    errorListener
-                )
-            ),
-            KlaxonDirectionsAPIResponseParser(),
-            errorListener
-        )
-
-        // Create a coroutine so we can invoke the suspend function Directions::getDirections
-        GlobalScope.launch {
-            val response = directions.getDirections(start, end, travelMode)
-            if (response != null) {
-                val startPoint = routeStartPoint(start, response)
-                val endPoint = routeEndPoint(end, response)
-                val routeBounds = routeRouteBounds(response)
-
-                val line = response.routes[0].overviewPolyline.points
-                val decoded = PolylineEncoding.decode(line)
-
-                /**
-                 * The com.google package contains two different representations of the
-                 * LatLng class. The decoded points must be converted to the representation
-                 * in com.google.android for PolylineOptions to accept them.
-                 */
-                val decodedAsGoodLatLng = decoded.map {
-                    LatLng(it.lat, it.lng)
-                }
-
-                /**
-                 * addPolyline throws an exception if it is not run on the Ui thread.
-                 */
-                activity.runOnUiThread {
-                    // Set the duration of the route
-                    val radioButtonId = "radio_" + travelMode.toLowerCase()
-                    val id = activity.resources.getIdentifier(radioButtonId, "id", activity.packageName)
-                    activity.findViewById<RadioButton>(id).apply {
-                        text = response.routes[0].legs[0].duration.text
-                    }
-                    polyLineOptions(startPoint, endPoint, routeBounds, decodedAsGoodLatLng)
-                }
+            }
+            else -> {
+                pathPolylines.add(createPathPolyline(
+                    start, end, travelMode
+                ))
             }
         }
     }
-
-    private fun polyLineOptions(startPoint: MarkerOptions, endPoint: MarkerOptions, routeBounds: LatLngBounds, decodedAsGoodLatLng: List<LatLng>) {
-        val polyOptions = PolylineOptions()
-            .color(COLOR_BLUE_ARGB.toInt())
-            .pattern(PATTERN_POLYGON_ALPHA)
-            .addAll(decodedAsGoodLatLng)
-        polyline = map.addPolyline(polyOptions)
-        begin = map.addMarker(startPoint)
-        dest = map.addMarker(endPoint)
-        map.moveCamera(
-            CameraUpdateFactory.newLatLngBounds(
-                routeBounds,
-                100 // padding around the route in pixels
-            )
-        )
+    public fun removeFromMap() {
+        pathPolylines.forEach{
+            it.removeFromMap()
+        }
+    }
+    public fun getPathPolylines(): List<PathPolyline> {
+        return this.pathPolylines
+    }
+    private fun isIndoorLocation(encodedLocation: String): Boolean {
+        return encodedLocation.startsWith(Constants.INDOOR_LOCATION_IDENTIFIER)
     }
 
-    private fun routeStartPoint(start: String, response: GoogleDirectionsAPIResponse): MarkerOptions {
-
-        return MarkerOptions().position(
-            LatLng(
-                response.routes[0].legs[0].startLocation.lat.toDouble(),
-                response.routes[0].legs[0].startLocation.lng.toDouble()
-            )
-        ).title(Helper.capitalizeWords(start)).snippet(startString)
-    }
-
-    private fun routeEndPoint(end: String, response: GoogleDirectionsAPIResponse): MarkerOptions {
-
-        return MarkerOptions().position(
-            LatLng(
-                response.routes[0].legs[0].endLocation.lat.toDouble(),
-                response.routes[0].legs[0].endLocation.lng.toDouble()
-            )
-        ).title(Helper.capitalizeWords(end)).snippet(destString)
-    }
-
-    private fun routeRouteBounds(response: GoogleDirectionsAPIResponse): LatLngBounds {
-
-        return LatLngBounds(
-            LatLng(
-                response.routes[0].bounds.southwest.lat.toDouble(),
-                response.routes[0].bounds.southwest.lng.toDouble()
-            ),
-            LatLng(
-                response.routes[0].bounds.northeast.lat.toDouble(),
-                response.routes[0].bounds.northeast.lng.toDouble()
-            )
-        )
-    }
-
-    private fun decodeLine(line: String): List<LatLng> {
-        /**
-         * The com.google package contains two different representations of the
-         * LatLng class. The decoded points must be converted to the representation
-         * in com.google.android for PolylineOptions to accept them.
-         */
-        val decoded = PolylineEncoding.decode(line)
-        return decoded.map {
-            LatLng(it.lat, it.lng)
+    suspend fun waitUntilCreated() {
+        this.pathPolylines.forEach {
+            it.waitUntilCreated()
         }
     }
 
-    private fun runAddPolyline(decodedAsGoodLatLng: List<LatLng>) {
-        /**
-         * addPolyline throws an exception if it is not run on the Ui thread.
-         */
-        activity.runOnUiThread {
-            polyline = map.addPolyline(
-                PolylineOptions().addAll(
-                    decodedAsGoodLatLng))
+    fun getDuration(): Int {
+        var duration = 0
+        this.pathPolylines.forEach {
+            duration += it.segment.getDuration()
         }
+        return duration
     }
+
+    private fun createPathPolyline(start: LocationMetadata, end: LocationMetadata, travelMode: String): PathPolyline {
+
+        val directions = giveMeAnOutdoorDirections()
+        val segmentArgs =
+            SegmentArgs(travelMode, buildingIndexSingleton, directions)
+
+        val firstSegment = createSegment(start.encoded, segmentArgs)
+        val secondSegment = createSegment(end.encoded, segmentArgs)
+        secondSegment.appendTo(firstSegment)
+
+        return PathPolyline(start.name, end.name, firstSegment)
+    }
+
+    private fun createSegment(location: String, args: SegmentArgs): Segment {
+        return if (isIndoorLocation(location))
+            IndoorSegment(location, args)
+        else OutdoorSegment(location, args)
+    }
+
+
+
 }
