+++ app/src/main/java/com/example/campusguide/directions/ShuttleBusHelpers.kt
@@ -1,0 +1,14 @@
+package com.example.campusguide.directions
+
+fun ShuttleBusStopSGW(): LocationMetadata {
+    return LocationMetadata(
+        name = "Henry F. Hall Building front doors, 1455 De Maisonneuve Blvd. W., SGW Campus",
+        encoded = "45.497163,-73.578535"
+    )
+}
+fun ShuttleBusStopLOYOLA(): LocationMetadata {
+    return LocationMetadata(
+        name = "Loyola Chapel, 7137 Sherbrooke St. W., Loyola Campus",
+        encoded = "45.458424,-73.638369"
+    )
+}
